drf/requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cx-Oracle==8.1.0
distlib==0.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
eventlet==0.30.2
filelock==3.0.12
greenlet==1.0.0
httpie==2.4.0
idna==2.10
isort==5.6.4
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.1.0
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pylint==2.6.0
PySocks==1.7.1
pytz==2020.5
rabbitmq==0.2.0
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
vine==5.0.0
virtualenv==20.3.1
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
wrapt==1.12.1
